requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai>=1.35.0
python-telegram-bot==20.6
requests==2.31.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pytz==2024.1




